Move package from .core.datagrip to .datagrip
