+++ jetbrains-ultimate/resources-201/META-INF/ext-datagrip.xml
@@ -9,7 +9,7 @@
     <extensions defaultExtensionNs="com.intellij.database">
         <connectionInterceptor implementation="software.aws.toolkits.jetbrains.services.rds.auth.IamAuth"/>
         <connectionInterceptor implementation="software.aws.toolkits.jetbrains.services.redshift.auth.IamAuth"/>
-        <connectionInterceptor implementation="software.aws.toolkits.jetbrains.core.datagrip.auth.SecretsManagerAuth"/>
+        <connectionInterceptor implementation="software.aws.toolkits.jetbrains.datagrip.auth.SecretsManagerAuth"/>
     </extensions>
     <actions>
     <group id="aws.toolkit.explorer.redshift.cluster" popup="true" compact="false">
