+++ jetbrains-ultimate/src-201/software/aws/toolkits/jetbrains/services/rds/auth/IamAuth.kt
@@ -22,7 +22,7 @@ import software.amazon.awssdk.regions.Region
 import software.aws.toolkits.core.utils.getLogger
 import software.aws.toolkits.core.utils.info
 import software.aws.toolkits.jetbrains.core.credentials.ConnectionSettings
-import software.aws.toolkits.jetbrains.core.datagrip.getAwsConnectionSettings
+import software.aws.toolkits.jetbrains.datagrip.getAwsConnectionSettings
 import software.aws.toolkits.jetbrains.utils.ApplicationThreadPoolScope
 import software.aws.toolkits.resources.message
 import java.time.Instant
