+++ jetbrains-ultimate/src-201/software/aws/toolkits/jetbrains/ui/AwsAuthWidget.kt
@@ -14,8 +14,8 @@ import com.intellij.uiDesigner.core.GridLayoutManager
 import com.intellij.util.text.nullize
 import org.jetbrains.annotations.TestOnly
 import software.aws.toolkits.jetbrains.core.credentials.CredentialManager
-import software.aws.toolkits.jetbrains.core.datagrip.CREDENTIAL_ID_PROPERTY
-import software.aws.toolkits.jetbrains.core.datagrip.REGION_ID_PROPERTY
+import software.aws.toolkits.jetbrains.datagrip.CREDENTIAL_ID_PROPERTY
+import software.aws.toolkits.jetbrains.datagrip.REGION_ID_PROPERTY
 import software.aws.toolkits.jetbrains.core.region.AwsRegionProvider
 import software.aws.toolkits.jetbrains.utils.ui.selected
 import software.aws.toolkits.resources.message
